+++ src/views/button.usage.js
@@ -12,6 +12,17 @@ export default (view, group) => ({
         {
             title: 'Variations',
             demo: group.map(name => `${name}:{ text: "${name}" }`)
+        },
+        {
+            title: 'Button as a link',
+            demo: {
+                view: 'button',
+                data: {
+                    text: 'Click me',
+                    href: '#url',
+                    external: true
+                }
+            }
         }
     ]
 });